refactor: move filter bar and related code into its own subfolder
