+++ src/app/common/filter-bar/ShortcutManager.ts
