+++ src/app/common/filter-bar/clamp.spec.ts
