+++ src/app/common/filter-bar/clamp.ts
