+++ src/app/common/filter-bar/normalizeFilterFields.ts
