+++ src/app/data-planes/views/DataPlaneListView.vue
@@ -51,7 +51,7 @@
                 @change="route.update"
               >
                 <template #toolbar>
-                  <KFilterBar
+                  <FilterBar
                     class="data-plane-proxy-filter"
                     :placeholder="`tag: 'kuma.io/protocol: http'`"
                     :query="route.params.query"
@@ -104,9 +104,10 @@
 import DataPlaneList from '../components/DataPlaneList.vue'
 import type { DataPlaneCollectionSource } from '../sources'
 import ErrorBlock from '@/app/common/ErrorBlock.vue'
-import KFilterBar from '@/app/common/KFilterBar.vue'
+import FilterBar from '@/app/common/filter-bar/FilterBar.vue'
 import SummaryView from '@/app/common/SummaryView.vue'
 import type { MeSource } from '@/app/me/sources'
+
 </script>
 
 <style lang="scss" scoped>
