+++ src/app/gateways/sources.ts
@@ -1,4 +1,5 @@
 import { DataSourceResponse } from '@/app/application/services/data-source/DataSourcePool'
+import { normalizeFilterFields } from '@/app/common/filter-bar/normalizeFilterFields'
 import type KumaApi from '@/services/kuma-api/KumaApi'
 import type {
   PaginatedApiListResponse as CollectionResponse,
@@ -7,7 +8,6 @@ import type {
   DataPlaneOverview as DataplaneOverview,
   MeshGatewayDataplane,
 } from '@/types/index.d'
-import { normalizeFilterFields } from '@/utilities/normalizeFilterFields'
 type DataplaneTypeParams = {
   type: 'all' | 'delegated' | 'builtin'
 }
